+++ CHANGELOG.md
@@ -8,6 +8,17 @@
 
 ---
 
+### 2025-10-27
+
+#### 🔧 Maintenance
+
+- run on current directory and modify setup.sh (`4911c83`) - Qwenzy
+
+**Total**: 1 commit(s)
+
+---
+
+
 ### 2025-10-26
 
 #### 🐛 Bug Fixes